+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Reminders
 
+## Unreleased
+
+- Added a "Not Equals" (not equal to) operation for the Level and iLevel conditions (#14)
+
 ## [v11.0.2](https://github.com/pcg79/Reminders-Addon/tree/v11.0.2) (2024-08-13)
 [Full Changelog](https://github.com/pcg79/Reminders-Addon/compare/v10.1.0..v11.0.2)